requirements.txt
Flask==3.0.3
Pillow==9.0.1
waitress==3.0.0
SQLite4==0.1.1